[WIP] Make MaterializeEncodingPass be backend specific.
