+++ compiler/src/iree/compiler/Codegen/VMVX/test/CMakeLists.txt
@@ -17,6 +17,7 @@ iree_lit_test_suite(
     "assign_constant_ordinals.mlir"
     "link_executables.mlir"
     "lower_linalg_microkernels.mlir"
+    "materialize_encoding.mlir"
   TOOLS
     FileCheck
     iree-opt
